refactor(fast_io): decompose io_uring_stub into mirrored module layout
The non-Linux io_uring fallback previously lived in a single 2118-line
file at crates/fast_io/src/io_uring_stub.rs, which is ~226% over the
650-line LoC cap and far harder to keep aligned with the real
crates/fast_io/src/io_uring/ directory.

Split the stub into per-concern submodules whose filenames mirror the
Linux side one-to-one:

- mod.rs (entry + public re-exports)
- buffer_ring.rs, cancel.rs, config.rs, disk_batch.rs
- file_factory.rs, file_reader.rs, file_writer.rs
- linkat.rs, linked_chain.rs, registered_buffers.rs
- renameat2.rs, send_zc.rs, session_pool.rs, shared_ring.rs
- socket_factory.rs, socket_reader.rs, socket_writer.rs (Unix-only)
- statx.rs
- tests.rs (cross-cutting fallback / parity tests)

lib.rs now points the gated module attribute at the new directory:

    #[path = "io_uring_stub/mod.rs"]
    pub mod io_uring;

Behaviour is unchanged: every stub entry point still reports
io::ErrorKind::Unsupported, and the standard-I/O fallbacks for the
file/socket factories still produce byte-identical output across the
Auto and Disabled policies. Each new submodule sits well under the
650-line cap (largest: tests.rs at 387 lines).
