+++ crates/fast_io/src/io_uring_stub/file_writer.rs
@@ -1,0 +1,89 @@
+//! Stub io_uring writer mirroring [`crate::io_uring::file_writer::IoUringWriter`].
+//! Creating always fails with `Unsupported`.
+
+use crate::io_uring_common::IoUringConfig;
+use crate::traits::FileWriter;
+use std::io::{self, Seek, SeekFrom, Write};
+use std::path::Path;
+
+/// Stub io_uring writer (not available on this platform).
+///
+/// Creating always fails with `Unsupported`.
+pub struct IoUringWriter {
+    _private: (),
+}
+
+impl IoUringWriter {
+    /// Always returns an `Unsupported` error on this platform.
+    pub fn create<P: AsRef<Path>>(_path: P, _config: &IoUringConfig) -> io::Result<Self> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+
+    /// Creates a file with preallocated space (always fails on this platform).
+    pub fn create_with_size<P: AsRef<Path>>(
+        _path: P,
+        _size: u64,
+        _config: &IoUringConfig,
+    ) -> io::Result<Self> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+
+    /// Writes data at the specified offset (always fails on this platform).
+    pub fn write_at(&mut self, _offset: u64, _buf: &[u8]) -> io::Result<usize> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+}
+
+impl Write for IoUringWriter {
+    fn write(&mut self, _buf: &[u8]) -> io::Result<usize> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+}
+
+impl Seek for IoUringWriter {
+    fn seek(&mut self, _pos: SeekFrom) -> io::Result<u64> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+}
+
+impl FileWriter for IoUringWriter {
+    fn bytes_written(&self) -> u64 {
+        0
+    }
+
+    fn sync(&mut self) -> io::Result<()> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+
+    fn preallocate(&mut self, _size: u64) -> io::Result<()> {
+        Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "io_uring is not available on this platform",
+        ))
+    }
+}
